update - course, lesson papers
